# internal/docs/swagger.yaml
definitions:
...
  github_com_umar5678_go-backend_internal_models.UserRole:
    enum:
    - rider
    - driver
    - admin
    - delivery_person
    - service_provider
    - handyman
    type: string
    x-enum-varnames:
    - RoleRider
    - RoleDriver
    - RoleAdmin
    - RoleDeliveryPerson
    - RoleServiceProvider
    - RoleHandyman
  github_com_umar5678_go-backend_internal_models.UserStatus:
    enum:
    - active
    - suspended
    - banned
    - pending_verification
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusSuspended
    - StatusBanned
    - StatusPendingVerification
...
  github_com_umar5678_go-backend_internal_modules_auth_dto.AuthResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
    type: object
  github_com_umar5678_go-backend_internal_modules_auth_dto.EmailLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_umar5678_go-backend_internal_modules_auth_dto.EmailSignupRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
    required:
    - email
    - name
    - password
    - role
    type: object
  github_com_umar5678_go-backend_internal_modules_auth_dto.PhoneLoginRequest:
    properties:
      phone:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
        enum:
        - rider
        - driver
    required:
    - phone
    - role
    type: object
  github_com_umar5678_go-backend_internal_modules_auth_dto.PhoneSignupRequest:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
        enum:
        - rider
        - driver
    required:
    - name
    - phone
    - role
    type: object
  github_com_umar5678_go-backend_internal_modules_auth_dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      name:
        type: string
      phone:
        type: string
      profilePhotoUrl:
        type: string
      role:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
      status:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserStatus'
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverDashboardResponse:
    properties:
      acceptanceRate:
        type: number
      completionRate:
        type: number
      profile:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse'
      rating:
        type: number
      status:
        type: string
      todayEarnings:
        type: number
      todayTrips:
        type: integer
      totalTrips:
        type: integer
      walletBalance:
        type: number
      weekEarnings:
        type: number
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse:
    properties:
      acceptanceRate:
        type: number
      cancellationRate:
        type: number
      createdAt:
        type: string
      currentLocation:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.LocationResponse'
      heading:
        type: integer
      id:
        type: string
      isVerified:
        type: boolean
      licenseNumber:
        type: string
      rating:
        type: number
      status:
        type: string
      totalEarnings:
        type: number
      totalTrips:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      userId:
        type: string
      vehicle:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.VehicleResponse'
      walletBalance:
        type: number
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.LocationResponse:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateDriverProfileRequest:
    properties:
      licenseNumber:
        maxLength: 100
        minLength: 5
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateLocationRequest:
    properties:
      heading:
        maximum: 360
        minimum: 0
        type: integer
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
    required:
    - latitude
    - longitude
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateStatusRequest:
    properties:
      status:
        enum:
        - offline
        - online
        type: string
    required:
    - status
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateVehicleRequest:
    properties:
      capacity:
        maximum: 20
        minimum: 1
        type: integer
      color:
        maxLength: 50
        minLength: 2
        type: string
      isActive:
        type: boolean
      licensePlate:
        maxLength: 50
        minLength: 2
        type: string
      make:
        maxLength: 100
        minLength: 2
        type: string
      model:
        maxLength: 100
        minLength: 2
        type: string
      vehicleTypeId:
        type: string
      year:
        minimum: 1900
        type: integer
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.VehicleResponse:
    properties:
      capacity:
        type: integer
      color:
        type: string
      createdAt:
        type: string
      driverId:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      licensePlate:
        type: string
      make:
        type: string
      model:
        type: string
      updatedAt:
        type: string
      vehicleType:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse'
      vehicleTypeId:
        type: string
      year:
        type: integer
    type: object
  github_com_umar5678_go-backend_internal_modules_drivers_dto.WalletResponse:
    properties:
      balance:
        type: number
      lastTransaction:
        type: string
      pendingAmount:
        type: number
      totalEarnings:
        type: number
    type: object
...
  github_com_umar5678_go-backend_internal_modules_vehicles_dto.VehicleTypeResponse:
    properties:
      baseFare:
        type: number
      bookingFee:
        type: number
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      displayName:
        type: string
      iconUrl:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      perKmRate:
        type: number
      perMinuteRate:
        type: number
    type: object
...
  github_com_umar5678_go-backend_internal_utils_response.ErrorDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Meta:
    properties:
      durationMs:
        type: number
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Response:
    properties:
      code:
        type: string
      data: {}
      errors:
        description: Use ErrorDetail
        items:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.ErrorDetail'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Meta'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready Go backend API
  title: supr booking server in go
  version: "1.0"
paths:
  /auth/email/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.EmailLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.AuthResponse'
              type: object
      summary: Login with email
      tags:
      - auth
  /auth/email/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Signup data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.EmailSignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.AuthResponse'
              type: object
      summary: Signup with email (other roles)
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/phone/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.PhoneLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.AuthResponse'
              type: object
      summary: Login with phone (riders/drivers)
      tags:
      - auth
  /auth/phone/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Signup data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.PhoneSignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.AuthResponse'
              type: object
      summary: Signup with phone (riders/drivers)
      tags:
      - auth
  /auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.AuthResponse'
              type: object
      summary: Refresh access token
      tags:
      - auth
  /drivers/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverDashboardResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get driver dashboard stats
      tags:
      - drivers
  /drivers/location:
    post:
      consumes:
      - application/json
      parameters:
      - description: Location data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Update driver location
      tags:
      - drivers
  /drivers/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get driver profile
      tags:
      - drivers
    put:
      consumes:
      - application/json
      parameters:
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateDriverProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update driver profile
      tags:
      - drivers
  /drivers/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.DriverProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update driver status (online/offline)
      tags:
      - drivers
  /drivers/vehicle:
    put:
      consumes:
      - application/json
      parameters:
      - description: Vehicle update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.UpdateVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.VehicleResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update vehicle information
      tags:
      - drivers
  /drivers/wallet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_drivers_dto.WalletResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get driver wallet balance
      tags:
      - drivers
...
swagger: "2.0"